add progression display
